Added sorting function

+++ HoldYourHorses/Views/Sticks/IndexVM.cs
@@ -2,14 +2,13 @@
 {
     public class IndexVM
     {
-        public decimal PrisMin { get; set; }
-        public decimal PrisMax { get; set; }
+        public int PrisMin { get; set; }
+        public int PrisMax { get; set; }
         public int HästkrafterMin { get; set; }
         public int HästkrafterMax { get; set; }
         public string[] Materialer { get; set; } 
         public string[] Typer { get; set; }
 
-    
 
     }
